pyproject.toml
[project]
name = "what-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "google-generativeai>=0.8.5",
    "imageio-ffmpeg>=0.6.0",
    "openai>=1.107.2",
    "pdfminer>=20191125",
    "pdfminer-six>=20250506",
    "pydub>=0.25.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "reportlab>=4.4.3",
    "requests>=2.32.5",
    "sniffio==1.3.0",
    "supabase>=2.18.1",
]
